Create mytest.cpp
It is used to check the functions compress_plwah and decompress_plwah. 
The input file is an index file of wah. And then it will compress and decompress it by the two functions above.
We just need to compare the result file with the original file.
If they are the same, the two function are correct.